Can place ships on gameboard
Gameboards are now able to place ships for the user player.
Need to make a way to create the enemy's gameboard and a function printEnemyBoard that would show limited information.
Ships are done except for isAlive function. Unknown if we should implement.
Still have to write the logic for the actual battleship game.