Add files via upload
Finalized v2.3 of the AI. I added methods to pre-existing classes. I've already merged any new changes in with my changes in these files. I've also added a new AIShipData class that acts as the memory of the AI.  Challenge difficulty has not been implemented yet in this version, but the rest have. We'll also need to adjust the AI class name in runBattleship to AIv2 but that should be easy.